Experimental arc fitter. Small API changes. Add another example.
The new arc fitter is not tested well enough to be enabled
by default, but it appears to reduce the size of G-Code by several
times. Use caution! This fitter will not be applied to ramps yet,
as that would need even more extensive code changes. This will
be fixed in future.

The new 'semi-safe Z' parameter has been added to set the height
that is guaranteed to be above the top of the material. Rapid
downward moves are supposed to be restricted to above that level.

A new half-bracket example has been added. Helical drilling will
use 2 arc commands instead of 4 for compactness of the resulting
G-Code.