Optimize drawing of complicated shapes.
Anti-aliasing and thick pens are switched off when panning the view. Zooming
is still rather slow, but faster than before. A future improvement would be
to use a timer to initially draw a simplified shape when zooming, then follow
up with the final shape after the user is likely done with zooming.