Apply "arcification" to linear ramps. Very risky!
Found a couple of bugs while working on this, and probably added a few
new ones! Checked a few outputs in CAMotics and by cursory inspection
in text editor, but haven't developed a proper battery of tests yet.

While the lines-to-arc feature is disabled by default, the changes may
also impact other parts of the code, so proceed very cautiously.