Add a very embarassing example for text embossing.
This is a good demonstration of why the pocket milling toolpath code needs
a lot more work. It reveals the general slowness of the lines to arcs
algorithm too. The example creates a pocket with a large number of
complicated islands. The current algorithm follows those meandering
paths multiple times. This is unnecessary and adds to both computation
time and milling time. A better way might be to hog out large open areas
with simple paths and then deal with remaining areas of high detail
separately.