More features for (still untested) drilling. An untested face-mill operation.
The face-mill operation effectively implements axis-aligned pockets,
except no islands are supported yet.

Also, found a bug in joinClosePaths which may cause some corners to be
errorneously milled away on certain shapes. It might have not been a
problem before, when it was only used for contour-aligned pocket milling,
but I'm not 100% sure, especially in context of islands. However, it
definitely behaves incorrectly when face-milling concave shapes. This
might not be a huge problem when only used for the intended purpose, i.e.
flattening the stock. However it currently limits the possibility of
using the face-mill operation for axis aligned pocket milling or perhaps
in some unusual side clamping scenarios.