Add cutting ops that add draft. Fix a few latent bugs.
Those might be useful for making patterns for sand casting. It's not
a very fast method to cut drafts, as it requires cutting in tiny layers,
but there is a big advantage of being able to use standard cutting bits.

In fact, this should also be usable with tapered cutters at greatly
increased layer height. Of course, in that case, the draft angle
setting must be equal to (half of?) the taper angle of the cutter.