Very risky changes to improve draft angle support.
Chances are this breaks everything, as I had to redesign large parts of the
code that produces the G-Code for 2D paths, as well as do some rather risky
improvements to the code that merges adjacent passes. This needs some serious
testing/validation.

The upside is that milling items with draft angle should be several times
faster. It is now possible to use holding tabs for outside cuts, but the
draft angle will only extend to the tab depth.  Fixing this is not trivial
and the practical impact shouldn't be too great. I'm not limiting the depth
of cut now, and that should be OK for wooden patterns and reasonable draft
angles, but this might be something to improve in future.

The improved path joining should also help with regular pockets. However,
it is very slow and clumsy and I intend to try to at least make it faster.