Add initial API support for user-defined tabs.
No GUI support yet. Also, the tab locations are specified on the target
(offset) polyline, not the original polyline. It's not possible to change
tab length per tab.

I probably have most of the geometry code needed to offset the tabs from
the original polylines, but it may require a bit more work.