New tab handling code. Improved rendering code.
The tab code should be able to handle more difficult cases:

* Multiple polygons caused as an inside cutout of a shape that has
  passages narrower than the tool bit (for auto-tabs it will create
  tabs for each cutout; for manual tabs it will assign the right tabs
  to their corresponding shapes)
* Widened cutouts (it now tries to clone the tabs onto each pass)

The auto-tab feature is still pretty basic and in more complicated cases
(gears or any other complex outlines, especially with widened contours)
manual tab placing will be necessary.

The rendering code is made a bit simpler/cleaner to prepare for future
changes. Much of the detail of handling tabs is shifted to the operation
handling code, and the rendering code is passed a simple array of
(depth, path). This might also allow a better preview of shapes with
draft angles and other 2.5D features.