Start moving CAM code to its own directory + slotting experiments.
This new code is trying to do something that might resemble trochoidal
milling, but really is interspersing the original path with circles
that expand it outwards. This might help with chip evacuation, but
does not guarantee good cutting behaviour. Also, it hasn't been tested
with an actual machine. It doesn't support tabs. It's just an experiment.