Trochoidal slot fixes.
Looks like the logic that interpolates between two points on the outside curve
can start picking pretty random points if curr is far away from prev. In those
cases, it helps to use another heuristic that doesn't suffer from this problem
as a fallback. In this case, a point on the outside curve that is closest to
the step2 point of a line segment connecting the previous and the
current-proposed point (i.e. the current-proposed point if step2 is 1, or the
earlier point on the line when it's less than 1, basically linear
interpolation).

I'm not sure if it will handle enough edge cases, but it at least handles
the ones I know.