Add a setting for speed vs bug-safety of Z moves.
The existing code was producing very risk-averse paths. Add an option to
restore the (mostly) old behaviour, but default to a riskier choice of
entering the already cut areas of stock using rapid moves. However, the
last 1mm (or whatever is set as a safe entry Z height) towards the top of
the uncut stock is still done slowly to prevent the hard impact of the
tool against the material.