Attempt to fix dogbones for the refine operation.
This hacky fix connects the dogbone circle with the parent
shape using a convex hull of two tool diameter circles,
half a tool diameter apart. The last part might not be
good enough, it may possibly be too far into the shape
(and collide with islands etc.) or too close. Should
do the trick for the less extreme of the cases.