Use stepover allowance for helical entry into inside corners.
The entry helix results in a round cut, so it will never be able to end up
in a sharp corner. Mitigate this by allowing a side entry from the helix
to the ultimate starting point if it's shorter than the stepover value,
as this should be pretty close to any other kind of stepover move.

Now, if the angle is more acute and the helix ends up too far away,
then a helical move cannot be used. I don't think I'm handling that
properly.