Fix wide contours - broken helical entry and incorrect depth-first order.
When tabs are used on wide contours, the offset contour lines cannot
be merged. Also, they should be cut in width-first order, not depth-first,
so that the main goal (improved chip evacuation due to a wider contour line)
is met.

I don't fully understand why the code in cutlayers_for_margin has
to put in the contours in reverse order - the offset=0 line before
all the offset=extra width ones. I suspect that this is related
to contour sorting in CutLayerTree, but this needs further
investigation.