New implementation of v-carving.
This version should work a little better for metal engraving with
many shallow layers - it avoids most rapid moves at the expense
of longer cutting paths.

At the moment pocketing wide areas is not supported. Also, the
regular preview does not show the correct rapid moves and can
get confused badly on some shapes. This is work in progress.