Remove toolpath.Toolpaths. Do some related refactoring to help with that.
The Toolpaths object hasn't been that useful so far and leading to
obscure errors and confusion, so it is now replaced by the consistent
use of Toolpath for single paths and a Python list when multiple paths
can be used. This led to some unexpected changes when constructing
CutLayer2D objects and their subpaths argument. Now CutLayer2D holds
a list of Toolpath segments. If there are no tabs, it's a list with
a single item.

I've removed mergeToolpaths, I have some doubts about whether that's safe
or not. Also, TrochoidalContour is an old class that hasn't been used
in a while and seems broken, so it's going, too. The Contour class
and its direct base TabbedOperation have been merged, because there are
no other "tabbed" operations existing or even planned in near future.