Fix a regression in trochoidal contours caused by a switch to Pyclipr.
The regression was caused by a difference in semantics of the add operator.
In Clipper1 the list returned was a plain Python list (IIRC) so the add
operator of a result and a slice did a concatenation.

Clipper2 returns a numpy array that interprets the add operator in the
same context as a matrix addition type operation. Hence, to restore the
previous behaviour, I had to convert the result to a Python list.