Initial implementation of auto-joining lines/arcs on DXF import.
This is mostly for FreeCAD users. FreeCAD exports sketches as discrete
lines/arcs and does not produce closed polylines, which makes things
a little inconvenient.

This implementation has some flaws: it won't bypass duplicate objects
objects when loading a new version of the DXF, because it tries to
do the comparison before the line-joining step. It also leaves the
join operation on the undo stack, which is harmless, and sometimes
even useful, but not totally expected.

Also, a crash bug in the equals operator for PathArc has been fixed.