Add max depth of cut per pass. Add some tooltips in cutter manager.
Also, fix (maybe) the issue of controls being wrongly grayed out when
editing an endmill.

The DOC per pass is meant for relieved and/or indexable cutters where
the total depth of cut in terms of clearance is larger than the length
of the cutting edge. It can also be used as a safety feature on hobby
machines that lack power or rigidity for high values of axial
engagement (Ap).